requirements/base.txt
Django>=3.0
pyjwt>=1.7.1